requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
PyMuPDF==1.23.6
pandas==2.1.1
scikit-learn==1.3.2
Pillow==10.1.0
google-generativeai==0.3.1
requests==2.31.0
python-dateutil==2.8.2